+++ src/screens/Dashboard/Profile/index.tsx
@@ -1,0 +1,17 @@
+import { ScreenTitle } from 'components/Common';
+import React, { FC } from 'react';
+import { StyleProps } from 'styles';
+
+type Props = StyleProps;
+
+export const DashboardProfileScreen: FC<Props> = () => {
+  return (
+    <>
+      <ScreenTitle title="Profile" />
+      {`Profile`}
+    </>
+  );
+};
+
+export type DashboardProfileScreenProps = Props;
+export default DashboardProfileScreen;